*.wks.in: Replace hardcoded values with variables
Replace the hardcoded values "openstlinux-weston" and "stm32mp"
with variables ${DISTRO} and ${MACHINE}. To get this working
wks file extension with wks.in.

As a result of this improvement the STM32MP BSP layer is capable
of generating WIC for any of the supported machines and for any
of the openembedded-core images, for example core-image-minimal,
core-image-weston, etc.

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>